requirements.txt
django==3.1.7
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
phonenumbers==8.13.9
pillow==9.5.0
python-resize-image==1.1.20